gameapi/app/requirements.txt
anyio==3.5.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
decorator==5.1.1
fastapi==0.71.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.8.1
mysql-connector-python==8.0.28
oauthlib==3.1.1
pbr==5.8.0
protobuf==3.19.3
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
starlette==0.17.1
Tempita==0.5.2
tweepy==4.4.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
zipp==3.6.0
